# docs/swagger.yaml
basePath: /
definitions:
  models.Entry:
    properties:
      cl:
        type: integer
      date:
        type: string
      datetime:
        description: 'TODO: convert from "2020-04-15 18:50:13" to ISO.'
        type: string
      email:
        type: string
      enheter:
        type: integer
      host:
        type: string
      id:
        type: integer
      ip:
        type: string
      lat:
        type: number
      likes:
        type: integer
      lon:
        type: number
      msg:
        type: string
      olsug:
        type: integer
      personal_secret:
        type: boolean
      place:
        type: string
      report:
        type: boolean
      secret:
        type: boolean
      sideKicks:
        items:
          $ref: '#/definitions/models.SideKick'
        type: array
      sig:
        type: string
      status:
        type: integer
      time:
        type: string
    type: object
  models.Member:
    properties:
      address:
        type: string
      address_url:
        type: string
      email:
        type: string
      history:
        type: string
      id:
        type: integer
      im:
        type: string
      is_valid:
        type: boolean
      name:
        type: string
      number:
        type: integer
      password:
        type: string
      password_classic:
        type: string
      password_classic_resetstring:
        type: string
      password_resetstring:
        type: string
      phone:
        type: string
      picture:
        type: string
      title:
        type: string
    type: object
  models.SideKick:
    properties:
      id:
        type: integer
      number:
        type: string
    type: object
  router.File:
    properties:
      filename:
        type: string
    type: object
  router.Mail:
    properties:
      from_email:
        type: string
      message:
        type: string
      title:
        type: string
      to_emails:
        items:
          type: string
        type: array
    type: object
  router.ReturnMail:
    properties:
      result:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 'Backend for sidan. Authentication: Visit /swagger-auth to login with
    Google and get your JWT token, then paste it in the Authorize dialog.'
  license:
    name: MIT
    url: http://opensource.org/licenses/MIT
  title: Sidan API
  version: "3.0"
paths:
  /db/entries:
    get:
      parameters:
      - default: 20
        description: Number of entries to return
        in: query
        name: take
        type: integer
      - default: 0
        description: Number of entries to skip
        in: query
        name: skip
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Entry'
            type: array
      summary: List all entries
      tags:
      - entries
    post:
      consumes:
      - application/json
      parameters:
      - description: Entry data
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/models.Entry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entry'
      summary: Create entry
      tags:
      - entries
  /db/entries/{id}:
    delete:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entry'
      security:
      - BearerAuth: []
      summary: Delete entry
      tags:
      - entries
    get:
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entry'
      summary: Get entry by ID
      tags:
      - entries
    put:
      consumes:
      - application/json
      parameters:
      - description: Entry ID
        in: path
        name: id
        required: true
        type: integer
      - description: Entry data
        in: body
        name: entry
        required: true
        schema:
          $ref: '#/definitions/models.Entry'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Entry'
      security:
      - BearerAuth: []
      summary: Update entry
      tags:
      - entries
  /db/members:
    get:
      description: Returns full member data with read:member scope, limited data otherwise
      parameters:
      - default: false
        description: Return only valid members
        in: query
        name: onlyValid
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Member'
            type: array
      summary: List all members
      tags:
      - members
    post:
      consumes:
      - application/json
      parameters:
      - description: Member data
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      security:
      - BearerAuth: []
      summary: Create member
      tags:
      - members
  /db/members/{id}:
    delete:
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      security:
      - BearerAuth: []
      summary: Delete member
      tags:
      - members
    get:
      description: Returns full member data with read:member scope, limited data otherwise
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      summary: Get member by ID
      tags:
      - members
    put:
      consumes:
      - application/json
      parameters:
      - description: Member ID
        in: path
        name: id
        required: true
        type: integer
      - description: Member data
        in: body
        name: member
        required: true
        schema:
          $ref: '#/definitions/models.Member'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Member'
      security:
      - BearerAuth: []
      summary: Update member
      tags:
      - members
  /file/image:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Image file
        in: formData
        name: data
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.File'
      security:
      - BearerAuth: []
      summary: Upload image
      tags:
      - files
  /mail:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email data
        in: body
        name: mail
        required: true
        schema:
          $ref: '#/definitions/router.Mail'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/router.ReturnMail'
      security:
      - BearerAuth: []
      summary: Send email
      tags:
      - mail
securityDefinitions:
  BearerAuth:
    description: JWT Bearer token. Visit /swagger-auth to obtain a token via Google
      OAuth2.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"